refactor(orchestrator): migrate from Node --test to Bun test
- Rename .test.cjs to .test.js (15 test files) and helpers.cjs to helpers.js
- Remove require('node:test') imports; use Bun's global describe/it/test
- Keep require('node:assert') (Bun supports node:assert natively)
- Update require('./helpers.cjs') references to require('./helpers.js')
- Stub process.exit in websearch tests (core.cjs output() calls
  process.exit(0) which kills Bun's in-process test runner)
- Delete scripts/run-tests.cjs custom test runner
- Update test script to "bun test tests/"
- Remove test:coverage script and c8 devDependency

590 tests pass across 15 files.
