+++ apps/orchestrator/tests/agent-frontmatter.test.js
@@ -8,7 +8,6 @@
  * - Spawn type consistency across workflows
  */
 
-const { test, describe } = require('node:test');
 const assert = require('node:assert');
 const fs = require('fs');
 const path = require('path');
