+++ apps/orchestrator/tests/config.test.js
@@ -7,12 +7,11 @@
  * Requirements: TEST-13
  */
 
-const { test, describe, beforeEach, afterEach } = require('node:test');
 const assert = require('node:assert');
 const fs = require('fs');
 const path = require('path');
 const os = require('os');
-const { runKataTools, createTempProject, cleanup } = require('./helpers.cjs');
+const { runKataTools, createTempProject, cleanup } = require('./helpers.js');
 
 // ─── helpers ──────────────────────────────────────────────────────────────────
