+++ apps/orchestrator/tests/milestone.test.js
@@ -2,11 +2,10 @@
  * Kata Tools Tests - Milestone
  */
 
-const { test, describe, beforeEach, afterEach } = require('node:test');
 const assert = require('node:assert');
 const fs = require('fs');
 const path = require('path');
-const { runKataTools, createTempProject, cleanup } = require('./helpers.cjs');
+const { runKataTools, createTempProject, cleanup } = require('./helpers.js');
 
 describe('milestone complete command', () => {
   let tmpDir;
