+++ apps/orchestrator/tests/verify.test.js
@@ -2,11 +2,10 @@
  * Kata Tools Tests - Verify
  */
 
-const { test, describe, beforeEach, afterEach } = require('node:test');
 const assert = require('node:assert');
 const fs = require('fs');
 const path = require('path');
-const { runKataTools, createTempProject, createTempGitProject, cleanup } = require('./helpers.cjs');
+const { runKataTools, createTempProject, createTempGitProject, cleanup } = require('./helpers.js');
 const { execSync } = require('child_process');
 
 // ─── helpers ──────────────────────────────────────────────────────────────────
